remove deprecated Encoding key from desktop file
The Encoding key is now deprecated by the FreeDesktop standard and
all strings are required to be encoded in UTF-8.
This desktop entry explicitly specifies an Encoding of UTF-8,
which is harmless but no longer necessary.

The desktop-file-validate took in the desktop-file-utils package is useful
for checking the syntax of desktop entries.

Refer to http://standards.freedesktop.org/desktop-entry-spec/1.0/apc.html
for details.